Refactor scripts directory
Create a GamePass.json and MegaList.json that will be pushed to git
and have ids for any games in game pass and MegaList.json will store
all known product ids ever scraped.

Create execute folder with files that are meant to be executed
using ts-node.

Create scrape folder with functions to scrape data.

Refactor utils to an understandable way