Change checkbox to edit any user data for a game
For a game, a user previously would only ever have one piece of
data for the game, being whether or not the user owned the game.
Now, I have created a checkbox that will allow for more data and
can easily be reused whenever I want to add another field.