Create files that tried to create title images
Basically, I wanted to create title images, for example for
rocket league have the Rocket League title as its own image.
The product did not have images for the exact title but it
seemed like it had 2 of the same images but 1 with the title.
So my plan was to diff the 2 images and that would leave me with
just the title. However, after using this for the rocket league
images I noticed that the alignment of the images was not 100%
aligned, and I am guessing this is the same for a lot of products.
So for now, gonna move on but wanted to still save my work since
I think it should still work and could be useful later.